Crate advanced_widget_impl

Structs

App 🔒

BlueSquare 🔒

A widget that renders a blue square.

BoxedSquares 🔒

A widget that contains a list of several different widgets.

Greeting 🔒

An ephemeral greeting widget.

RedSquare 🔒

A widget that renders a red square.

RightAlignedSquare 🔒

A widget that renders a green square aligned to the right of the area.

Timer 🔒

A timer widget that displays the elapsed time since the timer was started.

Functions

fill 🔒

Fill the area with the specified symbol and style.

main 🔒